# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Random Video Chat Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: string;
  username: string;
  gender: string;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  age?: number | null;
  interests?: string[];
  language: string;
  createdAt: string;
}

export type ProfileInputGender = typeof ProfileInputGender[keyof typeof ProfileInputGender];


export const ProfileInputGender = {
  male: 'male',
  female: 'female',
  other: 'other',
} as const;

export interface ProfileInput {
  /** @minLength 1 */
  username: string;
  gender: ProfileInputGender;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  age?: number | null;
  interests?: string[];
  language?: string;
}

export interface TranslationRequest {
  text: string;
  /** @nullable */
  sourceLang?: string | null;
  targetLang: string;
}

export interface TranslationResponse {
  translatedText: string;
  originalText: string;
  targetLang: string;
  /** @nullable */
  sourceLang?: string | null;
}

export interface ErrorResponse {
  error: string;
}