# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Random Video Chat Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: string;
  username: string;
  gender: string;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  age?: number | null;
  interests?: string[];
  language: string;
  createdAt: string;
}