# lib/api-zod/src/generated/types/profileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Random Video Chat Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileInputGender } from './profileInputGender';

export interface ProfileInput {
  /** @minLength 1 */
  username: string;
  gender: ProfileInputGender;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  age?: number | null;
  interests?: string[];
  language?: string;
}