# lib/api-zod/src/generated/types/translationResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Random Video Chat Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TranslationResponse {
  translatedText: string;
  originalText: string;
  targetLang: string;
  /** @nullable */
  sourceLang?: string | null;
}